Add Task type and pipeline configuration
The pipeline configuration is an important part of the deployment
process.

A single task is something that should be run with the ECS runTask
function.

At the moment I am finding the command from the name of the task, but
there may be built-in tasks, like update_services, that shouldn't be a
command. Perhaps we need a special keyword in the deployment to assign?

Basically we will need some logic (maybe a very large case statement)
that says when this keyword is discovered, do this thing.