Formulate structure
This creates a more defined structure for the pipeline. Using pipeline
configuration we can create more flexible tasks, and achieve most things
we need to for our deployment.

Within a pipeline key word, we have a slice of different "steps". These
steps are called the type of thing they wish to achieve: task, script or
service.

These types can be asserted by the name of the step, and then we can
translate different type parameters from inside these maps.

I've also removed the requirement to specify an environment. This makes
it more flexible while I'm still trying to construct the format of the
tool.