Use ConfigOptions type to define config
Rather than repeat the values in two separate types (`Config` and
`Environment`), instead create a new type called `ConfigOptions` which
holds all the different configuration options we wish to provide.

This means adding a new option we only have to add it to one place, and
it will automatically be configurable at the top level at the
environment level, although a bit of code would still be required to
merge the options in.