Start to pad out logic for updating a service
The `Create()` function will probably only really run a few times. The
important one is updating the service itself with a new task definition.

This commit pads out the way that it should run. Firstly, the service
step within run.go should be consistent with other steps, by using a
`Run()` function for the `ServiceStep` type.

This `Run()` function should then take care of the lifecycle of the
step. If the service exists, update it. If it doesn't, create it.

Eventually it will also handle updating configuration for a service
itself, and in cases where the service has to be recreated, handle the
lifecycle of safely creating a new service and destroying the old one.