Add logic to recreate services
If a service exists, but we wish to recreate it, then we can use the
`--recreate-services` flag to automatically recreate services in the
pipeline.

Reasons we may wish to recreate are that we've changed something about
the service that we can't update in place, for example the LaunchType or
load balancer settings. In this case, updates would still apply, but the
new configuration would not.

Perhaps in the future we could automatically apply configuration by
comparing the local configuration and what the service itself is
configured with, but being able to force recreation seems very useful by
itself.