Re-work client stubbing
There are times when we want to be able to pass in multiple responses
from the client, so this slightly reworks it so we have a generic mocked
client, which has multiple responses that we can configure how we need
to.

This should make it straight forward to add some tests.