Fix bug with loading configuration
`Pipeline` was specified in two places, where we should only specify it
in a single place.

This also fixes the way that ConfigOptions are loaded, since I think I
broke it when trying to unify configuration. This is because it expected
the ConfigOptions to be in an outer map.

We can use the `inline` flag when unmarshalling it as described in
https://godoc.org/gopkg.in/yaml.v2#Marshal

This ensures that `Options` loads everything from the root of the
configuration file.