Add expressions to config
I had a dream about this and it turns out my dreams are pretty good
ideas. This implements adding expressions to the configuration, using
a simple regex to allow putting either the environment, project_name or
tag within the configuration itself.

This will be pretty powerful when you want to come to name resources
since you can use top level configuration with expressions.