Improve pipeline syntax
I've always been a bit wary of the pipeline syntax, the odd way that
it's an array of a key with an inner map:

```
{
  [
    "script":
      {
        "inline": "foo bar"
      }
  ]
}
```

This updates the syntax so it's a flat array of maps:

```
{
  [
    "type": "script",
    "inline": "foo bar"
  ]
}
```

This makes the pipeline configuration easier to read, but it also makes
it easier to _load_, because it will automatically load it when we set
the loading mechanism to `inline`.

The drawback with this approach is that we cannot have duplicate keys...
but that I think is probably a good thing, but it may lead to obscure
errors when loading configuration.