Load tasks in the same way we load services
To keep things consistent, we can define one off tasks in the same way
we define services. I think this will make the pipeline easier to read.

It also has the added benefit of being able to run specific tasks one
off using the CLI, because we can select from a predefined list of
tasks.