Rename DockerStep to BuildStep
I want to change the logic so you do not define builds inline, rather
select a build in a similar way that we do for services and tasks.

Rather than call it "docker" and have them defined as "dockers", it
seems more apt to call the type "build" and define a series of "builds".