Begin to add basic logic for load balancer
Adding all the logic to allow creation of a load balancer is quite
complicated, and something I had previously resisted to doing. However,
if I want to do a very basic example of a web app running, then it's
something I need to do, and I would like to be able to do a full
end-to-end example of a simple app.

This adds the example configuration so I can start to feel out what it's
going to look like.