Continue to add load balancer logic
This updates the logic to only create the load balancer if it does not
already exist, the same with the target group.

I think eventually we'll need to be able to re-create the target group
when we recreate the service for safety and so this will need to be
broken out; but while I'm still padding it out this will do.